+++ src/debugAdapter/Common/Contracts.ts
