+++ src/debugAdapter/Common/processServiceFactory.ts
@@ -4,9 +4,9 @@
 'use strict';
 
 import { inject, injectable } from 'inversify';
-import { ProcessService } from '../../../common/process/proc';
-import { IBufferDecoder, IProcessService, IProcessServiceFactory } from '../../../common/process/types';
-import { IServiceContainer } from '../../../ioc/types';
+import { ProcessService } from '../../client/common/process/proc';
+import { IBufferDecoder, IProcessService, IProcessServiceFactory } from '../../client/common/process/types';
+import { IServiceContainer } from '../../client/ioc/types';
 
 @injectable()
 export class DebuggerProcessServiceFactory implements IProcessServiceFactory {
