+++ src/debugAdapter/Common/protocolLogger.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+'use strict';
+
 import { injectable } from 'inversify';
 import { Readable } from 'stream';
 import { Logger } from 'vscode-debugadapter';
