+++ src/debugAdapter/Common/protocolParser.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+'use strict';
+
 // tslint:disable:no-constant-condition no-typeof-undefined
 
 import { EventEmitter } from 'events';
