+++ src/debugAdapter/Common/protocolWriter.ts
