+++ src/debugAdapter/main.ts
@@ -17,15 +17,15 @@ import { Disposable } from 'vscode';
 import { DebugSession, ErrorDestination, Event, logger, OutputEvent, Response, TerminatedEvent } from 'vscode-debugadapter';
 import { LogLevel } from 'vscode-debugadapter/lib/logger';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { EXTENSION_ROOT_DIR } from '../../common/constants';
-import '../../common/extensions';
-import { isNotInstalledError } from '../../common/helpers';
-import { IFileSystem } from '../../common/platform/types';
-import { ICurrentProcess } from '../../common/types';
-import { createDeferred, Deferred, sleep } from '../../common/utils/async';
-import { noop } from '../../common/utils/misc';
-import { IServiceContainer } from '../../ioc/types';
-import { AttachRequestArguments, LaunchRequestArguments } from '../types';
+import { EXTENSION_ROOT_DIR } from '../client/common/constants';
+import '../client/common/extensions';
+import { isNotInstalledError } from '../client/common/helpers';
+import { IFileSystem } from '../client/common/platform/types';
+import { ICurrentProcess } from '../client/common/types';
+import { createDeferred, Deferred, sleep } from '../client/common/utils/async';
+import { noop } from '../client/common/utils/misc';
+import { AttachRequestArguments, LaunchRequestArguments } from '../client/debugger/types';
+import { IServiceContainer } from '../client/ioc/types';
 import { CreateAttachDebugClient, CreateLaunchDebugClient } from './DebugClients/DebugFactory';
 import { BaseDebugServer } from './DebugServers/BaseDebugServer';
 import { initializeIoc } from './serviceRegistry';
