+++ src/debugAdapter/serviceRegistry.ts
@@ -4,17 +4,17 @@
 'use strict';
 
 import { Container } from 'inversify';
-import { SocketServer } from '../../common/net/socket/socketServer';
-import { FileSystem } from '../../common/platform/fileSystem';
-import { PlatformService } from '../../common/platform/platformService';
-import { IFileSystem, IPlatformService } from '../../common/platform/types';
-import { CurrentProcess } from '../../common/process/currentProcess';
-import { BufferDecoder } from '../../common/process/decoder';
-import { IBufferDecoder, IProcessServiceFactory } from '../../common/process/types';
-import { ICurrentProcess, ISocketServer } from '../../common/types';
-import { ServiceContainer } from '../../ioc/container';
-import { ServiceManager } from '../../ioc/serviceManager';
-import { IServiceContainer, IServiceManager } from '../../ioc/types';
+import { SocketServer } from '../client/common/net/socket/socketServer';
+import { FileSystem } from '../client/common/platform/fileSystem';
+import { PlatformService } from '../client/common/platform/platformService';
+import { IFileSystem, IPlatformService } from '../client/common/platform/types';
+import { CurrentProcess } from '../client/common/process/currentProcess';
+import { BufferDecoder } from '../client/common/process/decoder';
+import { IBufferDecoder, IProcessServiceFactory } from '../client/common/process/types';
+import { ICurrentProcess, ISocketServer } from '../client/common/types';
+import { ServiceContainer } from '../client/ioc/container';
+import { ServiceManager } from '../client/ioc/serviceManager';
+import { IServiceContainer, IServiceManager } from '../client/ioc/types';
 import { DebugStreamProvider } from './Common/debugStreamProvider';
 import { DebuggerProcessServiceFactory } from './Common/processServiceFactory';
 import { ProtocolLogger } from './Common/protocolLogger';
